backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.12.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=7.4.0
pytest-asyncio>=0.21.0
hypothesis>=6.80.0
httpx>=0.25.0
pypdf>=4.0.0
# Additional dependencies for error handling and monitoring
psutil>=5.9.0  # System monitoring (optional)

# Gemini AI for document processing (NEW package)
google-genai>=0.2.0
